requirements.txt
wheel
asn1crypto
setuptools
pydub
scipy
pyworld
matplotlib
numpy
soundfile
bandmat
pysptk
nnmnkwii
librosa
wavenet_vocoder
textgrid
torchinfo